requirements.txt
alembic==1.16.4
asyncpg==0.30.0
bcrypt==4.0.1
fastapi~=0.116.1
fastapi_auth_middlewares==0.1.0
passlib==1.7.4
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
SQLAlchemy==2.0.43
uvicorn==0.35.0